hub/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
dagger-io==0.19.4
nats-py==2.7.2
anyio>=4.2.0
opentelemetry-exporter-otlp-proto-grpc>=1.38.0
celery==5.3.4
redis==5.0.1
flower==2.0.1

# CLI tools
click>=8.1.7
dateparser>=1.2.0
rich>=13.7.0
pytest>=7.4.0
